split test command host effect fixtures
