+++ docs/PHASE_PLAN.md
@@ -473,6 +473,9 @@ checked-in docs, tests, and commits only.
 - Direct `zen build.zen` validation tests now keep graph-only library source
   and typecheck cases in a focused child module, separating them from
   executable target and gated test dependency validation.
+- Test-command build graph host-effect tests now keep declared file-read
+  fallback accept/reject cases in a focused child module, leaving env-effect
+  ordering and unrelated executable skip coverage in the parent module.
 - Resolver-backed behavior impl metadata now builds restored impl-block tasks
   once and reuses them for both impl method signature restoration and omitted
   default-method synthesis. Impl-block declaration collection now also uses a
